backend/requirements.txt
# =============================================================================
# government-ai backend — Python dependencies (pinned)
# Install:  python -m venv .venv
#           .venv/Scripts/Activate.ps1      # Windows PowerShell
#           pip install --upgrade pip
#           pip install -r requirements.txt
# When adding a package: add ONE line name==version here, then pip install -r requirements.txt
# =============================================================================

absl-py==2.4.0
aenum==3.1.17
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.20.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
boto3==1.42.77
botocore==1.42.77
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
cobble==0.1.4
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
databricks-sdk==0.102.0
dataclasses-json==0.6.7
deepgram-sdk==3.11.0
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
ecdsa==0.19.2
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.9.0.post1
fastapi==0.115.0
fastuuid==0.14.0
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gepa==0.1.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.0.3
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
huey==2.6.0
huggingface_hub==1.8.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
jax==0.9.2
jaxlib==0.9.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
limits==5.8.0
litellm==1.82.6
llama-cpp-python==0.3.19
lxml==6.0.2
Mako==1.3.10
mammoth==1.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==5.25.0
numpy==2.1.1
ollama==0.3.3
openai==2.28.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opentelemetry-api==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
opt_einsum==3.4.0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pdf2image==1.17.0
pillow==10.4.0
prettytable==3.17.0
propcache==0.4.1
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
pyparsing==3.3.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
pyzbar==0.1.9
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
skops==0.13.0
slowapi==0.1.9
smmap==5.0.3
sniffio==1.3.1
sounddevice==0.5.5
SQLAlchemy==2.0.30
sqlparse==0.5.5
starlette==0.38.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.30.6
waitress==3.0.2
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.7
wrapt==1.17.3
yarl==1.23.0
zipp==3.23.0
# Advanced RAG (backend/backend/rag_engine): BM25 + Chroma + LangGraph; Jina/Qwen via llama-cpp-python
rank-bm25==0.2.2
chromadb==1.5.5
langgraph==1.1.3
langchain-text-splitters==1.1.1
langchain-core==1.2.20